requirements.txt
transformers==4.22.1
torch==1.12.1
numpy==1.23.1
scikit-learn==1.0.2
wandb==0.12.1
pandas==1.3.5
rich==10.12.0
hydra-core==1.1.1
omegaconf==2.1.1
torchvision==0.10.3

